+++ backend/services/emailServices.js
@@ -61,6 +61,37 @@ export const sendEmail = async (mailOpts) => {
 };
 
 /**
+ * Sends an email with the given options and retries the operation if it fails.
+ * Retries are done with an exponential backoff.
+ * @param {TMailOpts} mailOpts
+ * @param {number} [retries=1]
+ * @return {Promise<{ wasSuccessful: boolean }>} A promise that resolves to an object with a boolean indicating whether the operation was successful.
+ */
+export const sendEmailWithRetries = async (mailOpts, retries = 1) => {
+    try {
+        const { wasSuccessful } = await sendEmail(mailOpts)
+
+        if (!wasSuccessful) {
+            throw new Error("Failed to send email. Retrying...");
+        }
+
+        return { wasSuccessful: true };
+    } catch (error) {
+        if (retries <= 3) {
+            const randomNumMs = Math.floor(Math.random() * (5000 - 1000 + 1)) + 1000;
+            const waitTime = randomNumMs + (retries * 1_000);
+            console.log("Will wait for ", waitTime);
+            await new Promise((resolve) => setTimeout(resolve, waitTime));
+            return await sendEmailWithRetries(mailOpts, retries + 1);
+        }
+
+        console.error("Failed to send the email.")
+
+        return { wasSuccessful: false };
+    }
+};
+
+/**
  * Adds a user to our email list via the brevo api.
  * @param {string} email The user's email.
  * @param {string} clientUrl The url of the client side of the app. This is used to construct the unsubscribe link.
@@ -110,38 +141,6 @@ export const addUserToEmailList = async (email, clientUrl) => {
         return { wasSuccessful: true };
     } catch (error) {
         console.error(error);
-
-        return { wasSuccessful: false };
-    }
-};
-
-
-/**
- * Sends an email with the given options and retries the operation if it fails.
- * Retries are done with an exponential backoff.
- * @param {TMailOpts} mailOpts
- * @param {number} [retries=1]
- * @return {Promise<{ wasSuccessful: boolean }>} A promise that resolves to an object with a boolean indicating whether the operation was successful.
- */
-export const sendEmailWithRetries = async (mailOpts, retries = 1) => {
-    try {
-        const { wasSuccessful } = await sendEmail(mailOpts)
-
-        if (!wasSuccessful) {
-            throw new Error("Failed to send email. Retrying...");
-        }
-
-        return { wasSuccessful: true };
-    } catch (error) {
-        if (retries <= 3) {
-            const randomNumMs = Math.floor(Math.random() * (5000 - 1000 + 1)) + 1000;
-            const waitTime = randomNumMs + (retries * 1_000);
-            console.log("Will wait for ", waitTime);
-            await new Promise((resolve) => setTimeout(resolve, waitTime));
-            return await sendEmailWithRetries(mailOpts, retries + 1);
-        }
-
-        console.error("Failed to send the email.")
 
         return { wasSuccessful: false };
     }
